Updated to Rails 2.3.4, and added zipfile upload
The new Rails is frozen in vendor.

Zipfile uploading requires rubyzip, and it is run through the usual image
upload form. However, if the "image" is a zipfile, then all images in the root
of that zipfile are added, with the description and tags entered in the form,
and with sequential names.

It's fairly rudimentary at the moment. What's missing is
  * Support for a hierarchical zipfile
  * Support for other archive types
  * Automatic use of special text files in the archive (for descriptions, etc)

But it's ok for now. I'll deploy it.