Made some environment adjustments
These are to make sure RMagick loads from a gem in $HOME/gems, and that it can
be linked against an ImageMagick installation in $HOME/lib. Passenger (for
example) screws up the environment to ignore the user-defined environment
variables.

For some reason, require "RMagick" is needed, otherwise the gem custom loading
fails to find libMagickCore.so. I don't know why, because doing
  ldd $GEM_HOME/gems/rmagick-2.12.2/lib/rmagick/RMagick.so
shows that all is well.